requirements.txt
# Core Discord library with voice support (exact version from Dandelion)
py-cord[voice] @ https://github.com/solaluset/pycord/releases/download/v2.6.1/py_cord-2.6.1-py3-none-any.whl

# Audio processing and downloading (exact version from Dandelion)
yt-dlp==2025.6.30
PyNaCl>=1.5.0

# Database and ORM
SQLAlchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
alembic>=1.13.1

# Music service integrations  
spotipy>=2.23.0

# HTTP client
aiohttp>=3.9.1

# File handling and utilities
aiofiles>=23.2.1
python-dotenv>=1.0.0

# System monitoring
psutil>=5.9.0

# Performance (Linux/Mac only)
uvloop>=0.19.0; sys_platform != "win32"

# Logging and utilities
colorama>=0.4.6

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0